Fix: Resolve compiler warnings

+++ src/document.h
@@ -43,7 +43,7 @@ class document
 	// Basic accessors for now
 	int line_count() const;
 	size_t get_line_count() const;
-	std::shared_ptr<line> get_line(size_t index) const;
+	std::shared_ptr<line> get_line(int index) const;
 
 	int get_cursor_x() const;
 	int get_cursor_y() const;
